refactored the file structure of the api module

+++ api/lib/database/models/user.ts
@@ -1,21 +1,20 @@
-import {IAuth, IUser, IUserConfig} from "@herbivore/core/utils/interfaces";
-import {model, Model, Schema, SchemaTypes, Types} from "mongoose";
-import {encrypt, indexOf} from "@herbivore/core/utils/functions";
-import {DEFAULT_CONFIG} from "@herbivore/core/utils/constants";
+import {IAuth, IPassEntry, IUser, IUserConfig} from "@herbivore/core/utils/interfaces";
+import {model, Model, Schema, SchemaTypes, Types, Document} from "mongoose";
+import {encrypt, indexOf, DEFAULT_CONFIG} from "@herbivore/core/utils";
 import {ArgumentError, DatabaseError, LoginError} from "@herbivore/core/utils/errors"
 import {PasswordEntry} from "./password";
 
-interface IUserDoc extends IUser {
+interface IUserDoc extends IUser, Document {
 }
 
 interface IUserModel extends Model<IUserDoc> {
-	login(auth: IAuth): Promise<boolean>
+	login(auth: IAuth): Promise<IUserDoc>
 	checkPassword(id: string, password: string): Promise<boolean>
 	isUniqueUsername(username: string): Promise<boolean>
 	getSalt(id: string): Promise<string>
-	addPassword(userID: string, entryID: string): void
+	addPassword(userID: string, entryID: string): Promise<IUserDoc>
 	removePassword(userID: string, entryID: string): Promise<boolean>
-	updatePassword(userID: string, newPassword: string): Promise<any>
+	updatePassword(userID: string, newPassword: string): Promise<boolean>
 	getUserConfig(userID: string): Promise<IUserConfig>
 }
 
@@ -73,10 +72,14 @@ userSchema.pre(/find.*/, function() {
 userSchema.static('updatePassword', function(userID: string, newPassword: string) {
 	return User.findById(userID).then(user => {
 		if (!user)
-			throw new DatabaseError(`No user found with ID ${userID}`)
+			throw new DatabaseError(`Unknown userID ${userID}`)
 
 		user.password = encrypt(newPassword, user.salt)
-		return user.save()
+		return user.save().then(_ => {
+			return true
+		}).catch(_ => {
+			return false
+		})
 	})
 })
 
@@ -85,7 +88,7 @@ userSchema.static('login', async function(auth: IAuth): Promise<IUserDoc> {
 	let user = await this.findOne({username: auth.username}, 'id password salt')
 
 	if (!user)
-		throw new LoginError("Username not present")
+		throw new LoginError("Unknown username")
 
 	if (user.password !== (encrypt(auth.password, user.salt)))
 		throw new LoginError("Incorrect password")
@@ -95,10 +98,10 @@ userSchema.static('login', async function(auth: IAuth): Promise<IUserDoc> {
 
 userSchema.static('checkPassword', async function(id: string, password: string): Promise<boolean> {
 	// TODO: Determine what I actually need from the login
-	let user = await this.findById(id, 'id password salt')
+	let user: IUser | null = await User.findById(id, 'id password salt')
 
 	if (!user)
-		throw new LoginError("User not found")
+		throw new LoginError(`Unknown userID ${id}`)
 
 	if (user.password !== (encrypt(password, user.salt)))
 		throw new LoginError("Incorrect password")
@@ -106,37 +109,38 @@ userSchema.static('checkPassword', async function(id: string, password: string):
 	return true
 })
 
-userSchema.static('addPassword', function(userID: string, entryID: string) {
-	return this.findById(userID).then((user: IUserDoc) => {
+userSchema.static('addPassword', function(userID: string, entryID: string): Promise<IUserDoc> {
+	return User.findById(userID).then((user: IUserDoc | null) => {
+		if (!user)
+			throw new ArgumentError(`Unknown user with userID ${userID}`)
+
 		user.entries.push(entryID)
 		return user.save()
 	})
 })
 
 userSchema.static('removePassword', function(userID: string, entryID: string): Promise<boolean> {
-	return this.findById(userID).then((user: IUserDoc) => {
-		console.log(entryID)
-		console.log(user.entries)
-		/*user.passwords.findIndex((item, index) => {
-			if ((<IPasswordEntry>item).id == entryID)
-				return index
+	let removed: (IPassEntry | string)[]
+	return User.findById(userID).then((user: IUserDoc | null) => {
+		if (!user)
+			throw new ArgumentError(`Unknown user with userID ${userID}`)
 
-			return -1
-		})*/
 		let index = indexOf(user.entries, entryID, 'id')
-		console.log("FINAL INDEX: ", index)
 
 		if (index < 0)
-			throw new ArgumentError(`EntryID ${entryID} not present for UserID ${userID}`);
+			throw new ArgumentError(`Unknown entryID ${entryID} for userID ${userID}`);
 
-		const removed = user.entries.splice(index, 1);
+		removed = user.entries.splice(index, 1);
 		return user.save().then(_ => {
 			return PasswordEntry.findByIdAndDelete(entryID).then(_ => {
 				return true;
 			});
 		}).catch(async _ => {
-			user.entries.push(removed[0]);
-			await user.save()
+			if (removed.length > 0 && !user.entries.includes(removed[0])) {
+				user.entries.push(removed[0]);
+				await user.save()
+			}
+
 			return false
 		});
 	});
@@ -145,7 +149,7 @@ userSchema.static('removePassword', function(userID: string, entryID: string): P
 userSchema.static('getSalt', async function(id: any): Promise<string> {
 	return await User.findById(id).then(user => {
 		if (!user)
-			throw new DatabaseError(`Unable to find a user with id ${id}`)
+			throw new ArgumentError(`Unknown userID ${id}`)
 
 		return user.password
 	})
@@ -154,7 +158,7 @@ userSchema.static('getSalt', async function(id: any): Promise<string> {
 userSchema.static('getUserConfig', async function(userID: string): Promise<IUserConfig> {
 	return await User.findById(userID, 'config').then(user => {
 		if (!user)
-			throw new DatabaseError(`Unable to find user with id ${userID}`)
+			throw new ArgumentError(`Unknown userID ${userID}`)
 
 		return user.config
 	})
